Repin the DOM maps forward and record the associated-state transfer
The lane's three implementation commits are verified and accepted here. The
census over `dom`, `shadow-dom`, `custom-elements`, the template element and the
iframe element moves seven subtests across three files and nothing the other
way: `Node-isConnected.html: Test with iframes`, the shadow retargeting case
`test-003.html`, and both `template-content-hierarcy.html` subtests. Key sets
are identical before and after, so there is no pass-to-fail side to explain.

`dom_boa.json` and `dom_nodes_boa.json` are repinned from that census and
verified key-by-key against the maps they replace: no former pass demoted, no
key dropped, 410 promotions and 2808 keys added -- subtests the runner now
reaches rather than new tests, the arrears of the navigation and lifecycle
phases which had never been repinned. The broad `dom` guard is green for the
first time, and `dom/nodes` with it, so all fourteen testharness slices and both
reftest guards sit at `unexpected=0`.

`prepareNodeMove` stops snapshotting a move that stays inside one document. The
snapshot exists only to re-record `ownerDocument` on the far side of a
cross-document move and it is shadow-including, so an ordinary
`body.appendChild` of a deep nest of shadow hosts was paying a reflector walk
per node for nothing. `adoptNode`, which overrides the destination after the
fact and can turn a same-document move into a crossing one, takes the snapshot
itself when it does.

One redundant `u64` cast this lane had introduced in `dom/shadow.rs` is dropped.

The plan gains the "Associated-state transfer" phase with the census table, the
repin verification, the named regression manifest and the gates, and DOC_README
gains its entry. Both carry a correction the Ortet receipts forced: `article`
has moved to `0x86e02f7fcd1c5b04` on `main`, not in this lane -- an Ortet built
at the lane's own base renders the same digest -- and `frames` is bimodal at
this base in the same proportion either side, so no single digest can honestly
be recorded for it.

+++ components/script-runtime-api/dom/shadow.rs
@@ -258,7 +258,7 @@ impl<E: ScriptEngine> NativeFn<E> for TemplateHostOf {
         };
         let found = raw.with_dom(|dom| dom.template_host_of(raw.id()));
         match found {
-            Some(template) => reflect_pinned::<E>(cx, template.raw() as u64),
+            Some(template) => reflect_pinned::<E>(cx, template.raw()),
             None => Ok(cx.make_null()),
         }
     }
